requirements.txt
Django==4.1.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
